DBZ-4353 Clean up
